Move safe_division to utils
