+++ perun/check/average_amount_threshold.py
@@ -1,4 +1,4 @@
-""" The `Average Amount Threshold` groups all of the resources according to the unique identifier
+""" The `Average Amount Threshold` groups all the resources according to the unique identifier
 (uid; e.g. the function name) and then computes the averages of resource amounts as performance
 representants of baseline and target profiles. The computed averages are then compared (by division
 , and according to the set threshold the checker detects either ``Optimization`` or ``Degradation``
@@ -31,7 +31,7 @@ The example of output generated by `AAT` method is as follows::
 
 In the output above, we detected the ``Optimization`` between commits ``1eb3d6`` (target) and
 ``7813e3`` (baseline), where the average amount of running time for ``SLList_search`` function
-changed from about six seconds to hundred miliseconds. For these detected changes we report no
+changed from about six seconds to a hundred miliseconds. For these detected changes we report no
 confidence at all.
 """
 from __future__ import annotations
@@ -39,10 +39,10 @@ from __future__ import annotations
 from typing import Any, Iterable
 
 import perun.profile.convert as convert
-import perun.postprocess.regression_analysis.tools as tools
+import perun.utils.helpers as helpers
 
-from perun.utils.structs import DegradationInfo, PerformanceChange
 from perun.profile.factory import Profile
+from perun.utils.structs import DegradationInfo, PerformanceChange
 
 
 DEGRADATION_THRESHOLD = 2.0
@@ -84,7 +84,7 @@ def average_amount_threshold(
     for target_uid, target_average in target_averages.items():
         baseline_average = baseline_averages.get(target_uid, None)
         if baseline_average is not None:
-            difference_ratio = tools.safe_division(target_average, baseline_average)
+            difference_ratio = helpers.safe_division(target_average, baseline_average)
             if difference_ratio >= DEGRADATION_THRESHOLD:
                 change = PerformanceChange.Degradation
             elif 0.0 < difference_ratio <= OPTIMIZATION_THRESHOLD:
